+++ lib/src/controls/inputs/buttons/base.dart
@@ -1,4 +1,3 @@
-// import 'package:flutter/material.dart' as m;
 import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/foundation.dart';
 
@@ -109,10 +108,21 @@ abstract class BaseButton extends StatefulWidget {
     super.debugFillProperties(properties);
     properties
       ..add(FlagProperty('enabled', value: enabled, ifFalse: 'disabled'))
-      ..add(
-          DiagnosticsProperty<ButtonStyle>('style', style, defaultValue: null))
-      ..add(DiagnosticsProperty<FocusNode>('focusNode', focusNode,
-          defaultValue: null));
+      ..add(DiagnosticsProperty<ButtonStyle>(
+        'style',
+        style,
+        defaultValue: null,
+      ))
+      ..add(DiagnosticsProperty<FocusNode>(
+        'focusNode',
+        focusNode,
+        defaultValue: null,
+      ))
+      ..add(DiagnosticsProperty<bool>(
+        'autofocus',
+        autofocus,
+        defaultValue: false,
+      ));
   }
 }
 
@@ -203,7 +213,12 @@ class _BaseButtonState extends State<BaseButton> {
                           .copyWith(color: resolvedForegroundColor),
                     ),
                 textAlign: TextAlign.center,
-                child: widget.child,
+                // used to align the child without expanding the button
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [widget.child],
+                ),
               ),
             ),
           ),