+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [next]
 
 * fix: Pop the menu flyout before than calling the close callback ([#1009](https://github.com/bdlukaa/fluent_ui/issues/1009))
+* fix: Button child is now automatically centered ([#1011](https://github.com/bdlukaa/fluent_ui/issues/1011))
 
 ## 4.8.5